+++ Backend/src/main/java/pl/HomeworkJustClick/Backend/Services/EvaluationServiceImplement.java
@@ -1,8 +1,10 @@
 package pl.HomeworkJustClick.Backend.Services;
 
 import jakarta.persistence.EntityManager;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import pl.HomeworkJustClick.Backend.Entities.Assignment;
 import pl.HomeworkJustClick.Backend.Entities.Evaluation;
 import pl.HomeworkJustClick.Backend.Entities.Solution;
 import pl.HomeworkJustClick.Backend.Entities.User;
@@ -10,8 +12,10 @@ import pl.HomeworkJustClick.Backend.Repositories.EvaluationRepository;
 import pl.HomeworkJustClick.Backend.Repositories.SolutionRepository;
 import pl.HomeworkJustClick.Backend.Repositories.UserRepository;
 import pl.HomeworkJustClick.Backend.Responses.EvaluationResponse;
+import pl.HomeworkJustClick.Backend.Responses.SolutionResponse;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class EvaluationServiceImplement implements EvaluationService {
@@ -31,13 +35,8 @@ public class EvaluationServiceImplement implements EvaluationService {
     }
 
     @Override
-    public Evaluation getById(int id) {
-        if(evaluationRepository.findById(id).isPresent()){
-            return evaluationRepository.findById(id).get();
-        }
-        else {
-            return null;
-        }
+    public Optional<Evaluation> getById(int id) {
+        return evaluationRepository.findById(id);
     }
 
     @Override
@@ -48,20 +47,44 @@ public class EvaluationServiceImplement implements EvaluationService {
                 .result(evaluation.getResult())
                 .userId(evaluation.getUser().getId())
                 .solutionId(evaluation.getSolution().getId())
+                .groupId(evaluation.getGroup().getId())
                 .creationDatetime(evaluation.getCreationDatetime())
                 .lastModifiedDatetime(evaluation.getLastModifiedDatetime())
-                .comment(evaluation.getComment())
                 .grade(evaluation.getGrade())
                 .build();
     }
 
     @Override
+    @Transactional
+    public EvaluationResponse addWithUserAndSolution(Evaluation evaluation, int user_id, int solution_id) {
+        Optional<User> user = userRepository.findById(user_id);
+        Optional<Solution> solution = solutionRepository.findById(solution_id);
+        if(user.isPresent() && solution.isPresent()) {
+            evaluation.setSolution(solution.get());
+            evaluation.setUser(user.get());
+            evaluation.setGroup(solution.get().getGroup());
+            entityManager.persist(evaluation);
+            return EvaluationResponse.builder()
+                    .id(evaluation.getId())
+                    .result(evaluation.getResult())
+                    .userId(evaluation.getUser().getId())
+                    .solutionId(evaluation.getSolution().getId())
+                    .groupId(evaluation.getGroup().getId())
+                    .creationDatetime(evaluation.getCreationDatetime())
+                    .lastModifiedDatetime(evaluation.getLastModifiedDatetime())
+                    .grade(evaluation.getGrade())
+                    .build();
+        } else {
+            return EvaluationResponse.builder().build();
+        }
+    }
+
+    @Override
     public Boolean delete(int id) {
-        try{
+        if(evaluationRepository.existsById(id)){
             evaluationRepository.deleteById(id);
             return true;
-        }
-        catch (IllegalArgumentException e){
+        } else {
             return false;
         }
     }
@@ -75,7 +98,7 @@ public class EvaluationServiceImplement implements EvaluationService {
             return true;
         }
         else{
-            return null;
+            return false;
         }
     }
 
@@ -89,7 +112,7 @@ public class EvaluationServiceImplement implements EvaluationService {
             return true;
         }
         else{
-            return null;
+            return false;
         }
     }
 
@@ -103,22 +126,10 @@ public class EvaluationServiceImplement implements EvaluationService {
             return true;
         }
         else{
-            return null;
+            return false;
         }
     }
 
-    @Override
-    public Boolean changeCommentById(int id, String comment) {
-        if(evaluationRepository.findById(id).isPresent()){
-            Evaluation evaluation = evaluationRepository.findById(id).get();
-            evaluation.setComment(comment);
-            evaluationRepository.save(evaluation);
-            return true;
-        }
-        else{
-            return null;
-        }
-    }
 
     @Override
     public Boolean changeGradeById(int id, Double grade) {
@@ -129,7 +140,7 @@ public class EvaluationServiceImplement implements EvaluationService {
             return true;
         }
         else{
-            return null;
+            return false;
         }
     }
 }
