+++ Backend/src/main/java/pl/HomeworkJustClick/Backend/Services/FileServiceImplement.java
@@ -11,6 +11,7 @@ import pl.HomeworkJustClick.Backend.Repositories.SolutionRepository;
 import pl.HomeworkJustClick.Backend.Responses.FileResponse;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class FileServiceImplement implements FileService {
@@ -25,6 +26,7 @@ public class FileServiceImplement implements FileService {
     SolutionRepository solutionRepository;
 
     EntityManager entityManager;
+    public FileServiceImplement(EntityManager entityManager) {this.entityManager = entityManager;};
 
     @Override
     public List<File> getAll() {
@@ -32,12 +34,8 @@ public class FileServiceImplement implements FileService {
     }
 
     @Override
-    public File getById(int id) {
-        if (fileRepository.findById(id).isPresent()) {
-            return fileRepository.findById(id).get();
-        } else {
-            return null;
-        }
+    public Optional<File> getById(int id) {
+        return fileRepository.findById(id);
     }
 
     @Override
@@ -45,6 +43,7 @@ public class FileServiceImplement implements FileService {
     public FileResponse addWithAssignment (File file, int assignment_id) {
         if(assignmentRepository.findById(assignment_id).isPresent()) {
             file.setAssignment(assignmentRepository.findById(assignment_id).get());
+            file.setSolution(null);
             entityManager.persist(file);
             return FileResponse.builder().id(file.getId()).name(file.getName()).format(file.getFormat()).mongo_id(file.getMongo_id()).build();
         } else {
@@ -57,6 +56,7 @@ public class FileServiceImplement implements FileService {
     public FileResponse addWithSolution (File file, int solution_id) {
         if(solutionRepository.findById(solution_id).isPresent()) {
             file.setSolution(solutionRepository.findById(solution_id).get());
+            file.setAssignment(null);
             entityManager.persist(file);
             return FileResponse.builder().id(file.getId()).name(file.getName()).format(file.getFormat()).mongo_id(file.getMongo_id()).build();
         } else {
@@ -66,10 +66,10 @@ public class FileServiceImplement implements FileService {
 
     @Override
     public Boolean delete(int id){
-        try {
+        if(fileRepository.existsById(id)) {
             fileRepository.deleteById(id);
             return true;
-        } catch (IllegalArgumentException e) {
+        } else  {
             return false;
         }
     }
