+++ Backend/src/main/java/pl/HomeworkJustClick/Backend/Services/SolutionServiceImplement.java
@@ -1,17 +1,22 @@
 package pl.HomeworkJustClick.Backend.Services;
 
 import jakarta.persistence.EntityManager;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import pl.HomeworkJustClick.Backend.Entities.Assignment;
+import pl.HomeworkJustClick.Backend.Entities.Group;
 import pl.HomeworkJustClick.Backend.Entities.Solution;
 import pl.HomeworkJustClick.Backend.Entities.User;
 import pl.HomeworkJustClick.Backend.Repositories.AssignmentRepository;
 import pl.HomeworkJustClick.Backend.Repositories.SolutionRepository;
 import pl.HomeworkJustClick.Backend.Repositories.UserRepository;
+import pl.HomeworkJustClick.Backend.Responses.AssignmentResponse;
 import pl.HomeworkJustClick.Backend.Responses.SolutionResponse;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class SolutionServiceImplement implements SolutionService{
@@ -31,13 +36,8 @@ public class SolutionServiceImplement implements SolutionService{
     }
 
     @Override
-    public Solution getById(int id) {
-        if(solutionRepository.findById(id).isPresent()){
-            return solutionRepository.findById(id).get();
-        }
-        else {
-            return null;
-        }
+    public Optional<Solution> getById(int id) {
+        return solutionRepository.findById(id);
     }
 
     @Override
@@ -49,16 +49,40 @@ public class SolutionServiceImplement implements SolutionService{
                 .assignmentId(solution.getAssignment().getId())
                 .creationDateTime(solution.getCreationDatetime())
                 .lastModifiedDatetime(solution.getLastModifiedDatetime())
+                .comment(solution.getComment())
                 .build();
     }
 
     @Override
+    @Transactional
+    public SolutionResponse addWithUserAndAssignment(Solution solution, int user_id, int assignment_id) {
+        Optional<User> user = userRepository.findById(user_id);
+        Optional<Assignment> assignment = assignmentRepository.findById(assignment_id);
+        if(user.isPresent() && assignment.isPresent()) {
+            solution.setAssignment(assignment.get());
+            solution.setUser(user.get());
+            solution.setGroup(assignment.get().getGroup());
+            entityManager.persist(solution);
+            return SolutionResponse.builder()
+                    .id(solution.getId())
+                    .userId(user_id)
+                    .assignmentId(assignment_id)
+                    .groupId(assignment.get().getGroup().getId())
+                    .creationDateTime(solution.getCreationDatetime())
+                    .lastModifiedDatetime(solution.getLastModifiedDatetime())
+                    .comment(solution.getComment())
+                    .build();
+        } else {
+            return SolutionResponse.builder().build();
+        }
+    }
+
+    @Override
     public Boolean delete(int id) {
-        try {
+        if (solutionRepository.existsById(id)) {
             solutionRepository.deleteById(id);
             return true;
-        }
-        catch (IllegalArgumentException e){
+        } else {
             return false;
         }
     }
@@ -73,7 +97,7 @@ public class SolutionServiceImplement implements SolutionService{
             return true;
         }
         else {
-            return null;
+            return false;
         }
     }
 
@@ -87,7 +111,44 @@ public class SolutionServiceImplement implements SolutionService{
             return true;
         }
         else {
-            return null;
+            return false;
         }
     }
+
+    @Override
+    public List<SolutionResponse> getSolutionsByGroupId(int id) {
+        List<Solution> solutions = solutionRepository.getSolutionsByGroupId(id);
+        List<SolutionResponse> solutionResponses = new ArrayList<>();
+        for(Solution solution : solutions) {
+            solutionResponses.add(SolutionResponse.builder()
+                    .id(solution.getId())
+                    .userId(solution.getUser().getId())
+                    .assignmentId(solution.getAssignment().getId())
+                    .groupId(solution.getGroup().getId())
+                    .creationDateTime(solution.getCreationDatetime())
+                    .lastModifiedDatetime(solution.getLastModifiedDatetime())
+                    .comment(solution.getComment())
+                    .build());
+        }
+        return solutionResponses;
+    }
+
+    @Override
+    public List<SolutionResponse> getSolutionsByAssignmentId(int id) {
+        List<Solution> solutions = solutionRepository.getSolutionsByAssignmentId(id);
+        List<SolutionResponse> solutionResponses = new ArrayList<>();
+        for(Solution solution : solutions) {
+            solutionResponses.add(SolutionResponse.builder()
+                    .id(solution.getId())
+                    .userId(solution.getUser().getId())
+                    .assignmentId(solution.getAssignment().getId())
+                    .groupId(solution.getGroup().getId())
+                    .creationDateTime(solution.getCreationDatetime())
+                    .lastModifiedDatetime(solution.getLastModifiedDatetime())
+                    .comment(solution.getComment())
+                    .build());
+        }
+        return solutionResponses;
+    }
+
 }
